Provide necessary context.
Due to one word ("on") not being taken over from the original article,
the given construction "võimalik see allkirjastada" seemed strange,
because the direct object in singular has to be either in the Genitive
or Partitive case.  In the given example the direct object is in the
nominative case.

By adding "on" however, it becomes clear that the construction just uses
an unusual[1] word order.

[1]: unusual for me at the time of writing.